Consolidate related extension files
- Merge Color+toHex.swift into Color+Mix.swift (both extend Color/NativeColor)
- Merge Solar+SolarEvent.swift into Solar++.swift (mutually dependent)
- Organize consolidated files with MARK sections for better navigation
- Update Xcode project to reflect file changes
